Split stock league demo responsibilities (Closes #2109)
