+++ adl/src/demo/stock_league/shared.rs
@@ -1,0 +1,89 @@
+pub(super) const DEMO_NAME: &str = "demo-i-v090-stock-league-scaffold";
+pub(super) const INTEGRATION_DEMO_NAME: &str = "demo-j-v090-stock-league-recurring";
+pub(super) const EXTENSION_DEMO_NAME: &str = "demo-k-v090-stock-league-proof-expansion";
+
+pub(super) const RUN_ID: &str = "demo-i-stock-league-scaffold-run-001";
+pub(super) const INTEGRATION_RUN_ID: &str = "demo-j-stock-league-recurring-run-001";
+pub(super) const EXTENSION_RUN_ID: &str = "demo-k-stock-league-proof-expansion-run-001";
+pub(super) const SEASON_ID: &str = "season-001";
+pub(super) const FIXED_TIME: &str = "2026-04-17T00:00:00Z";
+pub(super) const DISCLAIMER: &str = "This is a paper-market simulation for demonstrating persistent agent identity and accountability. It is not financial advice, trading advice, or a real investment strategy.";
+pub(super) const FIXTURE_JSON: &str =
+    include_str!("../../../../demos/fixtures/stock_league/season_001_fixture.json");
+
+#[derive(Clone, Copy)]
+pub(super) struct AgentSpec {
+    pub(super) id: &'static str,
+    pub(super) display_name: &'static str,
+    pub(super) role: &'static str,
+    pub(super) primary_lens: &'static str,
+    pub(super) risk_tolerance: &'static str,
+    pub(super) tension: &'static str,
+    pub(super) forbidden_behavior: &'static str,
+}
+
+pub(super) const AGENTS: &[AgentSpec] = &[
+    AgentSpec {
+        id: "value_monk",
+        display_name: "The Value Monk",
+        role: "competing_agent",
+        primary_lens: "valuation and balance-sheet discipline",
+        risk_tolerance: "moderate",
+        tension: "may underperform when momentum dominates",
+        forbidden_behavior: "opening a paper position solely because price is rising",
+    },
+    AgentSpec {
+        id: "momentum_surfer",
+        display_name: "The Momentum Surfer",
+        role: "competing_agent",
+        primary_lens: "relative strength and regime change",
+        risk_tolerance: "moderate_high",
+        tension: "must exit when the trend breaks instead of narrating around losses",
+        forbidden_behavior: "holding a broken trend without naming the failed setup",
+    },
+    AgentSpec {
+        id: "contrarian_raccoon",
+        display_name: "The Contrarian Raccoon",
+        role: "competing_agent",
+        primary_lens: "overreaction and broken narratives",
+        risk_tolerance: "moderate",
+        tension: "must separate disliked from permanently impaired",
+        forbidden_behavior: "disagreeing with consensus without evidence",
+    },
+    AgentSpec {
+        id: "quality_gardener",
+        display_name: "The Quality Gardener",
+        role: "competing_agent",
+        primary_lens: "durable margins and patient compounding",
+        risk_tolerance: "moderate_low",
+        tension: "must avoid overpaying for excellent businesses",
+        forbidden_behavior: "ignoring valuation because a business is high quality",
+    },
+    AgentSpec {
+        id: "macro_weather_oracle",
+        display_name: "The Macro Weather Oracle",
+        role: "competing_agent",
+        primary_lens: "rates, inflation, sector rotation, and liquidity",
+        risk_tolerance: "moderate",
+        tension: "macro stories can explain too much after the fact",
+        forbidden_behavior: "claiming an outcome was obvious only after it happened",
+    },
+    AgentSpec {
+        id: "risk_goblin",
+        display_name: "The Risk Goblin",
+        role: "risk_reviewer",
+        primary_lens: "concentration, drawdown, liquidity, and unsupported confidence",
+        risk_tolerance: "low",
+        tension: "must be useful rather than reflexively fearful",
+        forbidden_behavior: "blocking every paper action without a concrete risk reason",
+    },
+    AgentSpec {
+        id: "archivist_referee",
+        display_name: "The Archivist Referee",
+        role: "referee",
+        primary_lens: "append-only records, hindsight checks, and identity drift",
+        risk_tolerance: "not_applicable",
+        tension: "must be boring, precise, and difficult to fool",
+        forbidden_behavior: "silently rewriting prior commitments",
+    },
+];
